# openAPI_schema.yaml
openapi: 3.0.3
info:
  title: Nyaya API
  version: 1.0.0
  description: |
    OpenAPI specification for the Nyaya quiz and gamification platform.
    Covers authentication, quiz play, progress tracking, leaderboards,
    and admin content management.
servers:
  - url: https://api.nyaya.app/v1
    description: Production
  - url: http://localhost:8080/api/v1
    description: Local
security:
  - bearerAuth: []
...
paths:
...
  /categories/{categoryId}:
    get:
      tags: [Categories]
      summary: Get category details
      parameters:
        - $ref: '#/components/parameters/CategoryId'
      responses:
        '200':
          description: Category details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
  /categories/{categoryId}/quizzes:
    get:
      tags: [Categories, Quizzes]
      summary: List quizzes for a category
      parameters:
        - $ref: '#/components/parameters/CategoryId'
      responses:
        '200':
          description: Quizzes in category
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Quiz'
  /quizzes:
    get:
      tags: [Quizzes]
      summary: List quizzes
      parameters:
        - in: query
          name: categoryId
          schema:
            type: integer
            format: int64
        - in: query
          name: difficulty
          schema:
            type: string
            enum: [easy, medium, hard]
        - in: query
          name: active
          schema:
            type: boolean
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Quiz list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedQuizzes'
  /quizzes/{quizId}:
    get:
      tags: [Quizzes]
      summary: Get quiz details
      parameters:
        - $ref: '#/components/parameters/QuizId'
      responses:
        '200':
          description: Quiz details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quiz'
  /quizzes/{quizId}/questions:
    get:
      tags: [Quizzes]
      summary: Get quiz questions for play
      description: Correct answers are intentionally excluded.
      parameters:
        - $ref: '#/components/parameters/QuizId'
      responses:
        '200':
          description: Quiz questions
          content:
            application/json:
              schema:
                type: object
                properties:
                  quizId:
                    type: integer
                    format: int64
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/PlayableQuestion'
...
  /admin/categories:
    post:
      tags: [Admin]
      summary: Create category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryRequest'
      responses:
        '201':
          description: Category created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
  /admin/categories/{categoryId}:
    patch:
      tags: [Admin]
      summary: Update category
      parameters:
        - $ref: '#/components/parameters/CategoryId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCategoryRequest'
      responses:
        '200':
          description: Category updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
    delete:
      tags: [Admin]
      summary: Deactivate category
      parameters:
        - $ref: '#/components/parameters/CategoryId'
      responses:
        '204':
          description: Category deactivated
  /admin/quizzes:
    post:
      tags: [Admin]
      summary: Create quiz
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateQuizRequest'
      responses:
        '201':
          description: Quiz created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quiz'
  /admin/quizzes/{quizId}:
    patch:
      tags: [Admin]
      summary: Update quiz
      parameters:
        - $ref: '#/components/parameters/QuizId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateQuizRequest'
      responses:
        '200':
          description: Quiz updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quiz'
    delete:
      tags: [Admin]
      summary: Deactivate quiz
      parameters:
        - $ref: '#/components/parameters/QuizId'
      responses:
        '204':
          description: Quiz deactivated
  /admin/questions:
    post:
      tags: [Admin]
      summary: Create question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateQuestionRequest'
      responses:
        '201':
          description: Question created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Question'
  /admin/questions/{questionId}:
    patch:
      tags: [Admin]
      summary: Update question
      parameters:
        - $ref: '#/components/parameters/QuestionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateQuestionRequest'
      responses:
        '200':
          description: Question updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Question'
    delete:
      tags: [Admin]
      summary: Deactivate question
      parameters:
        - $ref: '#/components/parameters/QuestionId'
      responses:
        '204':
          description: Question deactivated
  /admin/questions/{questionId}/options:
    post:
      tags: [Admin]
      summary: Add question option
      parameters:
        - $ref: '#/components/parameters/QuestionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateQuestionOptionRequest'
      responses:
        '201':
          description: Option created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionOption'
  /admin/options/{optionId}:
    patch:
      tags: [Admin]
      summary: Update question option
      parameters:
        - $ref: '#/components/parameters/OptionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateQuestionOptionRequest'
      responses:
        '200':
          description: Option updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionOption'
    delete:
      tags: [Admin]
      summary: Delete question option
      parameters:
        - $ref: '#/components/parameters/OptionId'
      responses:
        '204':
          description: Option deleted
  /admin/levels:
    post:
      tags: [Admin]
      summary: Create level
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLevelRequest'
      responses:
        '201':
          description: Level created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Level'
  /admin/levels/{levelId}:
    patch:
      tags: [Admin]
      summary: Update level
      parameters:
        - $ref: '#/components/parameters/LevelId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLevelRequest'
      responses:
        '200':
          description: Level updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Level'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CategoryId:
      in: path
      name: categoryId
      required: true
      schema:
        type: integer
        format: int64
    QuizId:
      in: path
      name: quizId
      required: true
      schema:
        type: integer
        format: int64
    QuestionId:
      in: path
      name: questionId
      required: true
      schema:
        type: integer
        format: int64
...
    OptionId:
      in: path
      name: optionId
      required: true
      schema:
        type: integer
        format: int64
    LevelId:
      in: path
      name: levelId
      required: true
      schema:
        type: integer
        format: int64
    Page:
      in: query
      name: page
      schema:
        type: integer
        minimum: 1
        default: 1
    Size:
      in: query
      name: size
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  schemas:
...
    Category:
      type: object
      required: [id, name, slug, isActive]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Quiz:
      type: object
      required: [id, categoryId, title, difficulty, totalQuestions, isActive]
      properties:
        id:
          type: integer
          format: int64
        categoryId:
          type: integer
          format: int64
        title:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        difficulty:
          type: string
          enum: [easy, medium, hard]
        totalQuestions:
          type: integer
        timeLimitSeconds:
          type: integer
          nullable: true
        passingScore:
          type: number
          format: float
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Question:
      type: object
      required: [id, quizId, questionText, questionType, pointsReward, displayOrder, isActive]
      properties:
        id:
          type: integer
          format: int64
        quizId:
          type: integer
          format: int64
        questionText:
          type: string
        questionType:
          type: string
          enum: [single_choice, multi_choice, true_false]
        explanation:
          type: string
          nullable: true
        difficulty:
          type: string
          enum: [easy, medium, hard]
        pointsReward:
          type: integer
        negativePoints:
          type: integer
          default: 0
        displayOrder:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    QuestionOption:
      type: object
      required: [id, questionId, optionText, isCorrect, displayOrder]
      properties:
        id:
          type: integer
          format: int64
        questionId:
          type: integer
          format: int64
        optionText:
          type: string
        isCorrect:
          type: boolean
        displayOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
    PlayableQuestion:
      type: object
      required: [id, questionText, questionType, pointsReward, displayOrder, options]
      properties:
        id:
          type: integer
          format: int64
        questionText:
          type: string
        questionType:
          type: string
          enum: [single_choice, multi_choice, true_false]
        explanation:
          type: string
          nullable: true
        pointsReward:
          type: integer
        negativePoints:
          type: integer
        displayOrder:
          type: integer
        options:
          type: array
          items:
            type: object
            required: [id, optionText, displayOrder]
            properties:
              id:
                type: integer
                format: int64
              optionText:
                type: string
              displayOrder:
                type: integer
...
    Level:
      type: object
      required: [id, code, name, minPoints, maxPoints]
      properties:
        id:
          type: integer
          format: int64
        code:
          type: string
        name:
          type: string
        minPoints:
          type: integer
        maxPoints:
          type: integer
        badgeIcon:
          type: string
          nullable: true
        rewardDescription:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CreateCategoryRequest:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        isActive:
          type: boolean
          default: true
    UpdateCategoryRequest:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        isActive:
          type: boolean
    CreateQuizRequest:
      type: object
      required: [categoryId, title, slug, difficulty, totalQuestions]
      properties:
        categoryId:
          type: integer
          format: int64
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        totalQuestions:
          type: integer
        timeLimitSeconds:
          type: integer
        passingScore:
          type: number
          format: float
        isActive:
          type: boolean
          default: true
    UpdateQuizRequest:
      type: object
      properties:
        categoryId:
          type: integer
          format: int64
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        totalQuestions:
          type: integer
        timeLimitSeconds:
          type: integer
        passingScore:
          type: number
          format: float
        isActive:
          type: boolean
    CreateQuestionRequest:
      type: object
      required: [quizId, questionText, questionType, pointsReward, displayOrder]
      properties:
        quizId:
          type: integer
          format: int64
        questionText:
          type: string
        questionType:
          type: string
          enum: [single_choice, multi_choice, true_false]
        explanation:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        pointsReward:
          type: integer
        negativePoints:
          type: integer
        displayOrder:
          type: integer
        isActive:
          type: boolean
          default: true
    UpdateQuestionRequest:
      type: object
      properties:
        questionText:
          type: string
        questionType:
          type: string
          enum: [single_choice, multi_choice, true_false]
        explanation:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        pointsReward:
          type: integer
        negativePoints:
          type: integer
        displayOrder:
          type: integer
        isActive:
          type: boolean
    CreateQuestionOptionRequest:
      type: object
      required: [optionText, isCorrect, displayOrder]
      properties:
        optionText:
          type: string
        isCorrect:
          type: boolean
        displayOrder:
          type: integer
    UpdateQuestionOptionRequest:
      type: object
      properties:
        optionText:
          type: string
        isCorrect:
          type: boolean
        displayOrder:
          type: integer
    CreateLevelRequest:
      type: object
      required: [code, name, minPoints, maxPoints]
      properties:
        code:
          type: string
        name:
          type: string
        minPoints:
          type: integer
        maxPoints:
          type: integer
        badgeIcon:
          type: string
        rewardDescription:
          type: string
    UpdateLevelRequest:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        minPoints:
          type: integer
        maxPoints:
          type: integer
        badgeIcon:
          type: string
        rewardDescription:
          type: string
    PaginatedQuizzes:
      type: object
      properties:
        page:
          type: integer
        size:
          type: integer
        total:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/Quiz'
...